chore: renamed clients
