+++ src/client/clients/payments/open-api-definition.json
